Move shared account defaults from account to AccountDefaultsProvider
