+++ app-common/src/main/kotlin/net/thunderbird/app/common/account/CommonAccountDefaultsProvider.kt
@@ -3,11 +3,6 @@ package net.thunderbird.app.common.account
 import app.k9mail.core.featureflag.FeatureFlagProvider
 import app.k9mail.core.featureflag.toFeatureFlagKey
 import app.k9mail.legacy.account.Account
-import app.k9mail.legacy.account.Account.Companion.DEFAULT_SORT_ASCENDING
-import app.k9mail.legacy.account.Account.Companion.DEFAULT_SORT_TYPE
-import app.k9mail.legacy.account.Account.Companion.DEFAULT_SYNC_INTERVAL
-import app.k9mail.legacy.account.Account.Companion.NO_OPENPGP_KEY
-import app.k9mail.legacy.account.Account.Companion.UNASSIGNED_ACCOUNT_NUMBER
 import app.k9mail.legacy.account.Account.Expunge
 import app.k9mail.legacy.account.Account.FolderMode
 import app.k9mail.legacy.account.Account.ShowPictures
@@ -23,7 +18,12 @@ import app.k9mail.legacy.account.AccountDefaultsProvider.Companion.DEFAULT_QUOTE
 import app.k9mail.legacy.account.AccountDefaultsProvider.Companion.DEFAULT_REMOTE_SEARCH_NUM_RESULTS
 import app.k9mail.legacy.account.AccountDefaultsProvider.Companion.DEFAULT_REPLY_AFTER_QUOTE
 import app.k9mail.legacy.account.AccountDefaultsProvider.Companion.DEFAULT_RINGTONE_URI
+import app.k9mail.legacy.account.AccountDefaultsProvider.Companion.DEFAULT_SORT_ASCENDING
+import app.k9mail.legacy.account.AccountDefaultsProvider.Companion.DEFAULT_SORT_TYPE
 import app.k9mail.legacy.account.AccountDefaultsProvider.Companion.DEFAULT_STRIP_SIGNATURE
+import app.k9mail.legacy.account.AccountDefaultsProvider.Companion.DEFAULT_SYNC_INTERVAL
+import app.k9mail.legacy.account.AccountDefaultsProvider.Companion.NO_OPENPGP_KEY
+import app.k9mail.legacy.account.AccountDefaultsProvider.Companion.UNASSIGNED_ACCOUNT_NUMBER
 import app.k9mail.legacy.account.Identity
 import app.k9mail.legacy.notification.NotificationLight
 import app.k9mail.legacy.notification.NotificationSettings
