+++ legacy/core/src/main/java/com/fsck/k9/preferences/GeneralSettingsDescriptions.java
@@ -10,8 +10,8 @@ import java.util.TreeMap;
 import android.content.Context;
 
 import app.k9mail.feature.telemetry.api.TelemetryManager;
-import app.k9mail.legacy.account.Account;
 import app.k9mail.legacy.account.Account.SortType;
+import app.k9mail.legacy.account.AccountDefaultsProvider;
 import app.k9mail.legacy.di.DI;
 import app.k9mail.legacy.preferences.AppTheme;
 import app.k9mail.legacy.preferences.SubTheme;
@@ -178,10 +178,10 @@ class GeneralSettingsDescriptions {
             new V(102, new BooleanSetting(true))
         ));
         s.put("sortTypeEnum", Settings.versions(
-                new V(10, new EnumSetting<>(SortType.class, Account.DEFAULT_SORT_TYPE))
+                new V(10, new EnumSetting<>(SortType.class, AccountDefaultsProvider.getDEFAULT_SORT_TYPE()))
         ));
         s.put("sortAscending", Settings.versions(
-                new V(10, new BooleanSetting(Account.DEFAULT_SORT_ASCENDING))
+                new V(10, new BooleanSetting(AccountDefaultsProvider.DEFAULT_SORT_ASCENDING))
         ));
         s.put("theme", Settings.versions(
                 new V(1, new LegacyThemeSetting(AppTheme.LIGHT)),
